Hackathon FarmaVita/requirements.txt
# Core
boto3>=1.34.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Data processing
pyarrow>=14.0.0
pyspark>=3.5.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# ML tracking
mlflow>=2.10.0
xgboost>=2.0.0
optuna>=3.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.27.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
mangum>=0.17.0

# GenAI / RAG
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-aws>=0.1.0
faiss-cpu>=1.7.4

# Frontend (nice to have)
streamlit>=1.30.0

# Linting (dev)
ruff>=0.2.0
